music/research/beat_framework/requirements.txt
# Beat Framework — Python Dependencies
# ======================================
# Install with: pip install -r requirements.txt

# ── Core (required) ───────────────────────────────────────────
pyyaml>=6.0            # Config file parsing

# ── MIDI (strongly recommended) ──────────────────────────────
mido>=1.3.0            # MIDI read/write (pure Python fallback built-in if absent)
pretty_midi>=0.2.10    # Higher-level MIDI analysis (optional but useful)

# ── Data Collection ───────────────────────────────────────────
requests>=2.31.0       # HTTP requests (Billboard scraping, general use)
beautifulsoup4>=4.12.0 # HTML parsing for Billboard scraper
spotipy>=2.23.0        # Spotify Web API client
pylast>=5.2.0          # Last.fm API client

# ── Audio Rendering ───────────────────────────────────────────
pyfluidsynth>=1.3.3    # Python bindings for FluidSynth WAV rendering
                       # Also requires system FluidSynth:
                       #   macOS:  brew install fluid-synth
                       #   Ubuntu: sudo apt install fluidsynth libfluidsynth-dev

# ── ML Generation (optional, heavy) ──────────────────────────
# Uncomment to enable Magenta DrumsRNN + GrooVAE humanization:
# magenta>=2.1.4
# note-seq>=0.0.5
# tensorflow>=2.12.0   # Required by Magenta (CPU version works fine)

# ── Development / Analysis (optional) ────────────────────────
numpy>=1.24.0          # Numerical analysis
scipy>=1.11.0          # Statistical functions
matplotlib>=3.7.0      # Visualizing drum patterns
